extract recording controls component
